using objects instead of simple marks to mark cells
Each cell is now an object with a mark = 0 and a number ascending from 0>8.
When a player mark is added to the object, the number value of the cell is
added to the player's 'played cells' array.

TODO > need to make code to check the played cells array against the possible
winning combinations that are stored in another array.  I'll use 'split' and
'.contains' to compare against each group of numbers