display section finished?
display section SHOULD display the current state of the game when used,
but seems to be getting overridden by the prompts....  So.... Maybe we
move forward with the input forms, then we won't need the prompts and we
can see if that was what was actually causing the problem.  Also need to
consider whether to use a modal or just text above the game.  I think
I'm going to go with text above the game board.